# Add GF16 AVX2, AVX512 and SSSE3