--- a/_gen/gf16.go
+++ b/_gen/gf16.go
@@ -16,8 +16,14 @@ type table256 struct {
 	Lo, Hi reg.VecVirtual
 }
 
+// table128 contains memory pointers to tables
+type table128 struct {
+	Lo, Hi Op
+}
+
 type gf16ctx struct {
-	clrMask reg.VecVirtual
+	clrMask    reg.VecVirtual
+	clrMask128 reg.VecVirtual
 }
 
 func genGF16() {
@@ -156,6 +162,160 @@ func genGF16() {
 		VZEROUPPER()
 		RET()
 	}
+	// SSSE3:
+	{
+		TEXT("ifftDIT2_ssse3", attr.NOSPLIT, fmt.Sprintf("func(x, y []byte, table  *[8*16]uint8)"))
+		tablePtr := Load(Param("table"), GP64())
+		tables := [4]table128{}
+		for i, t := range tables {
+			// We almost have enough space for all tables.
+			if i > 2 {
+				t.Lo, t.Hi = Mem{Base: tablePtr, Disp: i * 16}, Mem{Base: tablePtr, Disp: i*16 + 16*4}
+			} else {
+				t.Lo, t.Hi = XMM(), XMM()
+				MOVUPS(Mem{Base: tablePtr, Disp: i * 16}, t.Lo)
+				MOVUPS(Mem{Base: tablePtr, Disp: i*16 + 16*4}, t.Hi)
+			}
+			tables[i] = t
+		}
+		// Generate mask
+		zero := XMM()
+		XORPS(zero, zero) // Zero, so bytes will be copied.
+		fifteen, mask := GP64(), XMM()
+		MOVQ(U32(0xf), fifteen)
+		MOVQ(fifteen, mask)
+		PSHUFB(zero, mask)
+		ctx.clrMask128 = mask
+
+		bytes := Load(Param("x").Len(), GP64())
+		x := Load(Param("x").Base(), GP64())
+		y := Load(Param("y").Base(), GP64())
+
+		Label("loop")
+		for i := 0; i < 2; i++ {
+			xLo, xHi, yLo, yHi := XMM(), XMM(), XMM(), XMM()
+			MOVUPS(Mem{Base: x, Disp: i*16 + 0}, xLo)
+			MOVUPS(Mem{Base: x, Disp: i*16 + 32}, xHi)
+			MOVUPS(Mem{Base: y, Disp: i*16 + 0}, yLo)
+			MOVUPS(Mem{Base: y, Disp: i*16 + 32}, yHi)
+			PXOR(xLo, yLo)
+			PXOR(xHi, yHi)
+			MOVUPS(yLo, Mem{Base: y, Disp: i*16 + 0})
+			MOVUPS(yHi, Mem{Base: y, Disp: i*16 + 32})
+			leoMulAdd128(ctx, xLo, xHi, yLo, yHi, tables)
+			MOVUPS(xLo, Mem{Base: x, Disp: i*16 + 0})
+			MOVUPS(xHi, Mem{Base: x, Disp: i*16 + 32})
+		}
+		ADDQ(U8(64), x)
+		ADDQ(U8(64), y)
+		SUBQ(U8(64), bytes)
+		JNZ(LabelRef("loop"))
+
+		RET()
+	}
+	{
+		TEXT("fftDIT2_ssse3", attr.NOSPLIT, fmt.Sprintf("func(x, y []byte, table  *[8*16]uint8)"))
+		tablePtr := Load(Param("table"), GP64())
+		tables := [4]table128{}
+		for i, t := range tables {
+			// We almost have enough space for all tables.
+			if i > 2 {
+				t.Lo, t.Hi = Mem{Base: tablePtr, Disp: i * 16}, Mem{Base: tablePtr, Disp: i*16 + 16*4}
+			} else {
+				t.Lo, t.Hi = XMM(), XMM()
+				MOVUPS(Mem{Base: tablePtr, Disp: i * 16}, t.Lo)
+				MOVUPS(Mem{Base: tablePtr, Disp: i*16 + 16*4}, t.Hi)
+			}
+			tables[i] = t
+		}
+		// Generate mask
+		zero := XMM()
+		XORPS(zero, zero) // Zero, so bytes will be copied.
+		fifteen, mask := GP64(), XMM()
+		MOVQ(U32(0xf), fifteen)
+		MOVQ(fifteen, mask)
+		PSHUFB(zero, mask)
+		ctx.clrMask128 = mask
+
+		bytes := Load(Param("x").Len(), GP64())
+		x := Load(Param("x").Base(), GP64())
+		y := Load(Param("y").Base(), GP64())
+
+		Label("loop")
+		for i := 0; i < 2; i++ {
+			xLo, xHi, yLo, yHi := XMM(), XMM(), XMM(), XMM()
+			MOVUPS(Mem{Base: y, Disp: i*16 + 0}, yLo)
+			MOVUPS(Mem{Base: y, Disp: i*16 + 32}, yHi)
+
+			prodLo, prodHi := leoMul128(ctx, yLo, yHi, tables)
+
+			MOVUPS(Mem{Base: x, Disp: i*16 + 0}, xLo)
+			MOVUPS(Mem{Base: x, Disp: i*16 + 32}, xHi)
+			PXOR(prodLo, xLo)
+			PXOR(prodHi, xHi)
+			MOVUPS(xLo, Mem{Base: x, Disp: i*16 + 0})
+			MOVUPS(xHi, Mem{Base: x, Disp: i*16 + 32})
+
+			PXOR(xLo, yLo)
+			PXOR(xHi, yHi)
+			MOVUPS(yLo, Mem{Base: y, Disp: i*16 + 0})
+			MOVUPS(yHi, Mem{Base: y, Disp: i*16 + 32})
+
+		}
+
+		ADDQ(U8(64), x)
+		ADDQ(U8(64), y)
+		SUBQ(U8(64), bytes)
+		JNZ(LabelRef("loop"))
+
+		RET()
+	}
+	{
+		TEXT("mulgf16_ssse3", attr.NOSPLIT, fmt.Sprintf("func(x, y []byte, table  *[8*16]uint8)"))
+		tablePtr := Load(Param("table"), GP64())
+		tables := [4]table128{}
+		for i, t := range tables {
+			// We have enough space for all tables.
+			if i > 3 {
+				t.Lo, t.Hi = Mem{Base: tablePtr, Disp: i * 16}, Mem{Base: tablePtr, Disp: i*16 + 16*4}
+			} else {
+				t.Lo, t.Hi = XMM(), XMM()
+				MOVUPS(Mem{Base: tablePtr, Disp: i * 16}, t.Lo)
+				MOVUPS(Mem{Base: tablePtr, Disp: i*16 + 16*4}, t.Hi)
+			}
+			tables[i] = t
+		}
+		bytes := Load(Param("x").Len(), GP64())
+		x := Load(Param("x").Base(), GP64())
+		y := Load(Param("y").Base(), GP64())
+		// Generate mask
+		zero := XMM()
+		XORPS(zero, zero) // Zero, so bytes will be copied.
+		fifteen, mask := GP64(), XMM()
+		MOVQ(U32(0xf), fifteen)
+		MOVQ(fifteen, mask)
+		PSHUFB(zero, mask)
+		ctx.clrMask128 = mask
+
+		Label("loop")
+		for i := 0; i < 2; i++ {
+			dataLo, dataHi := XMM(), XMM()
+			MOVUPS(Mem{Base: y, Disp: i*16 + 0}, dataLo)
+			MOVUPS(Mem{Base: y, Disp: i*16 + 32}, dataHi)
+
+			prodLo, prodHi := leoMul128(ctx, dataLo, dataHi, tables)
+			MOVUPS(prodLo, Mem{Base: x, Disp: i*16 + 0})
+			MOVUPS(prodHi, Mem{Base: x, Disp: i*16 + 32})
+		}
+
+		ADDQ(U8(64), x)
+		ADDQ(U8(64), y)
+		SUBQ(U8(64), bytes)
+		JNZ(LabelRef("loop"))
+
+		RET()
+	}
+
 }
 
 // xLo, xHi updated
@@ -195,5 +355,56 @@ func leoMul256(ctx gf16ctx, lo, hi reg.VecVirtual, table [4]table256) (prodLo, p
 	VPSHUFB(data1, table[3].Hi, tmpHi)
 	VPXOR(prodLo, tmpLo, prodLo)
 	VPXOR(prodHi, tmpHi, prodHi)
+	return
+}
+
+func leoMulAdd128(ctx gf16ctx, xLo, xHi, yLo, yHi reg.VecVirtual, table [4]table128) {
+	prodLo, prodHi := leoMul128(ctx, yLo, yHi, table)
+	PXOR(prodLo, xLo)
+	PXOR(prodHi, xHi)
+}
+
+// leoMul128 lo, hi preseved (but likely will take extra regs to reuse)
+func leoMul128(ctx gf16ctx, lo, hi reg.VecVirtual, table [4]table128) (prodLo, prodHi reg.VecVirtual) {
+	data0, data1 := XMM(), XMM()
+	MOVAPS(lo, data1)
+	PSRLQ(U8(4), data1) // data1 = lo >> 4
+	MOVAPS(lo, data0)
+	PAND(ctx.clrMask128, data0) // data0 = lo&0xf
+	PAND(ctx.clrMask128, data1) // data 1 = data1 &0xf
+	prodLo, prodHi = XMM(), XMM()
+	MOVUPS(table[0].Lo, prodLo)
+	MOVUPS(table[0].Hi, prodHi)
+	PSHUFB(data0, prodLo)
+	PSHUFB(data0, prodHi)
+	tmpLo, tmpHi := XMM(), XMM()
+	MOVUPS(table[1].Lo, tmpLo)
+	MOVUPS(table[1].Hi, tmpHi)
+	PSHUFB(data1, tmpLo)
+	PSHUFB(data1, tmpHi)
+	PXOR(tmpLo, prodLo)
+	PXOR(tmpHi, prodHi)
+
+	// Now process high
+	data0, data1 = XMM(), XMM() // Realloc to break dep
+	MOVAPS(hi, data0)
+	MOVAPS(hi, data1)
+	PAND(ctx.clrMask128, data0)
+	PSRLQ(U8(4), data1)
+	PAND(ctx.clrMask128, data1)
+
+	tmpLo, tmpHi = XMM(), XMM() // Realloc to break dep
+	MOVUPS(table[2].Lo, tmpLo)
+	MOVUPS(table[2].Hi, tmpHi)
+	PSHUFB(data0, tmpLo)
+	PSHUFB(data0, tmpHi)
+	PXOR(tmpLo, prodLo)
+	PXOR(tmpHi, prodHi)
+	MOVUPS(table[3].Lo, tmpLo)
+	MOVUPS(table[3].Hi, tmpHi)
+	PSHUFB(data1, tmpLo)
+	PSHUFB(data1, tmpHi)
+	PXOR(tmpLo, prodLo)
+	PXOR(tmpHi, prodHi)
 	return
 }