… galois_amd64.go
… galois_arm64.go
… galois_gen_amd64.go
… galois_gen_amd64.s
… galois_noasm.go
… galois_ppc64le.go
… leopard.go
… reedsolomon_test.go