Without/with AVX2:

```
BenchmarkEncode500x200x10K-32    	      30	  38367143 ns/op	 186.83 MB/s	 5255168 B/op	     513 allocs/op
BenchmarkEncode500x200x10K-32                950           1297672 ns/op        5523.74 MB/s        5560 B/op          1 allocs/op

BenchmarkReconstruct500x200x10K-32    	      97	  11080478 ns/op	  64.69 MB/s	 1073190 B/op	    1025 allocs/op
BenchmarkReconstruct500x200x10K-32           207           5747499 ns/op        1247.15 MB/s       50798 B/op          5 allocs/op
```

With SSSE3:
```
BenchmarkEncode500x200x10K-32                        417           2897345 ns/op        2473.99 MB/s       12636 B/op          2 allocs/op
BenchmarkVerify500x200x10000-32                      351           3503805 ns/op        2045.78 MB/s     2306301 B/op        225 allocs/op
BenchmarkReconstruct500x200x10K-32                   123           9728689 ns/op         736.79 MB/s       85473 B/op          9 allocs/op
BenchmarkReconstructData500x200x10K-32               122           9707279 ns/op         738.41 MB/s       86174 B/op          9 allocs/op
```

Pure Go after improvements:
```
BenchmarkEncode500x200x10K-32                         51          22357147 ns/op         320.61 MB/s      103147 B/op         11 allocs/op
BenchmarkVerify500x200x10000-32                       49          22739224 ns/op         315.23 MB/s     2173788 B/op        212 allocs/op
BenchmarkReconstruct500x200x10K-32                    19          57535779 ns/op         124.58 MB/s      553199 B/op         54 allocs/op
BenchmarkReconstructData500x200x10K-32                19          57603268 ns/op         124.44 MB/s      553199 B/op         54 allocs/op
```